+++ src/v1/rest/users/mod.rs
@@ -2,10 +2,6 @@
 //!
 //! <https://developers.google.com/gmail/api/reference/rest/v1/users>
 
-mod types;
-#[doc(inline)]
-pub use types::*;
-
 pub mod get_profile;
 pub mod stop;
 pub mod watch;